Tweak URL patterns a bit; SO-style canonization
All patterns for "slug" are changed from .*? to .+? because an empty slug does not make sense.

The blog post URLconf now allows a partially malformed URL to be redirected to the correct one as long as it contains a valid primary key value.

This provides a nice feature: a blog post can be referenced with a much shorter URL.